In the Sedov solver we force a time value of t=0 to be a very small number
This prevents errors in the common case where a user is trying to compare a code output with the solver across a range of times that include t = 0.
We choose a value of t=0 to be the fourth-root of the smallest possible number in Python. This is a small number (~1.2e-77) but not the smallest as this allowsroom for some mathematical manipulation without running into overflow errors.