Resolve "Sedov solver crashes at t=0"
* Bump version number to 1.7.3

* ake the Cog 1, 2, 7, 8, 9, 11, 13, 17 and 21 solvers return NaN vales when t=0

* Make Mader solver return a NaN array at t=0

* SuOlsen solver now returns NaN at t = 0

* Sedov solver now returns NaN for t = 0

* Bump version number to 1.7.2

* In the Sedov solver we force a time value of t=0 to be a very small number

This prevents errors in the common case where a user is trying to compare a code output with the solver across a range of times that include t = 0.
We choose a value of t=0 to be the fourth-root of the smallest possible number in Python. This is a small number (~1.2e-77) but not the smallest as this allowsroom for some mathematical manipulation without running into overflow errors.